Update Sourcefile to reflect design changes. Fix attempting to copy build/scripts/spm.sh regardless of its existance. Prepare for upcoming build_type configuration.

+++ bootstrap.sh
@@ -44,6 +44,6 @@ fi
 $TUP upd;
 
 if [[ "$?" == "0" ]]; then
-	cp -vfu build/scripts/spm.sh bin/spm
+	[[ -s "build/scripts/spm.sh" ]] && cp -vfu build/scripts/spm.sh bin/spm
 	echo "export PATH=$(pwd -P)/bin:\$PATH" > scripts/spm
 fi
